pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Taylor-Jayde  Blackstone <t.blackstone@inspyre.tech>"]
description = "A fully-featured, extensible bot for Discord."
name = "DiPH"
version = "1.0-alpha1"
[tool.poetry.dependencies]
"discord.py" = "*"
inspy_logger = "2.0.0a6"
requests = "*"
python = "^3.8"
flask = "^1.1.2"
[tool.poetry.dev-dependencies]
